Refactoring WIP, repaired saving colors bug

+++ ResourceMaker/Features/ResourceFileWindow/ResourceFile.cs
@@ -31,10 +31,10 @@ namespace Sklodowski.ResourceMaker.Features.ResourceFileWindow
         {
             writer.Write(BitmapFileName);
             writer.Write(HasTransparentColor ? 1 : 0);
-            writer.Write(HasTransparentColor ? TransparentColor.R : 0);
-            writer.Write(HasTransparentColor ? TransparentColor.G : 0);
-            writer.Write(HasTransparentColor ? TransparentColor.B : 0);
-            writer.Write(HasTransparentColor ? TransparentColor.A : 0);
+            writer.Write((byte)(HasTransparentColor ? TransparentColor.R : 0));
+            writer.Write((byte)(HasTransparentColor ? TransparentColor.G : 0));
+            writer.Write((byte)(HasTransparentColor ? TransparentColor.B : 0));
+            writer.Write((byte)(HasTransparentColor ? TransparentColor.A : 0));
 
             var layers = GetAllLayersNames();
             writer.Write(layers.Count);
